==> code/analyses/Visium/08_LC_pseudobulkDE.R <==
################################################################
# LC analyses: pseudobulked differential expression (DE) testing
# Lukas Weber, Oct 2022
################################################################

# adapting code by Leonardo Collado-Torres, Abby Spangler, and Sowmya Parthiban from:
# https://github.com/LieberInstitute/spatialDLPFC


# module load conda_R/4.2
# Rscript filename.R

# file location:
# /dcs04/lieber/lcolladotor/pilotLC_LIBD001/locus-c/


library(SpatialExperiment)
library(here)
library(scater)
library(scran)
library(limma)
library(dplyr)
library(tidyr)
library(tibble)
library(ggplot2)
library(ggnewscale)
library(ggrepel)
library(ComplexHeatmap)
library(viridisLite)


# directory to save plots
dir_plots <- here("plots", "Visium", "08_pseudobulkDE")

# directory to save outputs
dir_outputs <- here("outputs", "Visium", "08_pseudobulkDE")


# ---------
# load data
# ---------

# load saved SPE object from previous script

fn_spe <- here("processed_data", "SPE", "LC_filtered.rds")
spe <- readRDS(fn_spe)

dim(spe)

table(colData(spe)$sample_id)

sample_ids <- levels(colData(spe)$sample_id)
sample_ids


# ----------------
# pseudobulk spots
# ----------------

# pseudobulk spots within manually annotated LC regions and non-LC regions by sample

table(colData(spe)$sample_id, colData(spe)$annot_region)

annot_region_fctr <- factor(as.numeric(colData(spe)$annot_region), 
                            labels = c("non", "LC"))
table(annot_region_fctr)

ids <- DataFrame(
  annot_region_pseudo = annot_region_fctr, 
  sample_id_pseudo = colData(spe)$sample_id
)

table(ids$sample_id_pseudo, ids$annot_region_pseudo)

# pseudobulking
spe_pseudo <- aggregateAcrossCells(spe, ids)
dim(spe_pseudo)

# add levels to colData and column names
levs <- paste(
  colData(spe_pseudo)$annot_region_pseudo, 
  colData(spe_pseudo)$sample_id_pseudo, 
  sep = "."
)
colData(spe_pseudo)$level_id <- levs
colnames(spe_pseudo) <- levs

# check
dim(spe_pseudo)
head(colData(spe_pseudo), 3)
counts(spe_pseudo)[1:3, ]


# ---------
# logcounts
# ---------

# calculate logcounts for pseudobulked data
# using library size scale factors

spe_pseudo <- computeLibraryFactors(spe_pseudo)
spe_pseudo <- logNormCounts(spe_pseudo)


# ---------
# filtering
# ---------

# no additional filtering needed
# low-expressed genes were already filtered out previously


# -----------------------
# pseudobulked DE testing
# -----------------------

# define model: LC vs. non-LC regions with blocks by sample
model_formula <- ~annot_region_pseudo
model_matrix <- model.matrix(model_formula, data = colData(spe_pseudo))

model_formula
model_matrix

# calculate intra-block correlation using limma
corfit <- duplicateCorrelation(
  logcounts(spe_pseudo), 
  model_matrix, 
  block = colData(spe_pseudo)$sample_id_pseudo
)
corfit$consensus.correlation


# calculate DE tests per gene using limma
mat <- logcounts(spe_pseudo)

res <- eBayes(
  lmFit(
    mat, 
    design = model_matrix, 
    block = colData(spe_pseudo)$sample_id_pseudo, 
    correlation = corfit$consensus.correlation
  )
)


# extract p-values
p_vals <- res$p.value[, "annot_region_pseudoLC"]

# calculate adjusted p-values (FDRs)
fdrs <- p.adjust(p_vals, method = "fdr")


# store corresponding gene names for convenience later
stopifnot(length(fdrs) == length(rowData(spe_pseudo)$gene_id))
stopifnot(all(names(fdrs) == rowData(spe_pseudo)$gene_id))

fdrs_gene_ids <- rowData(spe_pseudo)$gene_id
fdrs_gene_names <- rowData(spe_pseudo)$gene_name

names(fdrs) <- fdrs_gene_names


# -------
# results
# -------

# number of significant DE genes
table(fdrs <= 0.05)
table(fdrs <= 1e-2)
table(fdrs <= 1e-3)

# top genes
sort(fdrs[fdrs <= 1e-3])


# check: plot most significant DE gene

# most significant DE gene
which.min(fdrs)
most_sig <- names(which.min(fdrs))

# using spot-level (not pseudobulked) SPE object for plotting
df <- as.data.frame(cbind(colData(spe), spatialCoords(spe)))
ix_most_sig <- which(rowData(spe)$gene_name == most_sig)
df$most_sig <- counts(spe)[ix_most_sig, ]

ggplot(df, aes_string(x = "pxl_col_in_fullres", y = "pxl_row_in_fullres", 
                      color = "most_sig")) + 
  facet_wrap(~ sample_id, nrow = 2, scales = "free") + 
  geom_point(size = 0.1) + 
  scale_color_gradient(low = "gray80", high = "red", trans = "sqrt", 
                       name = "counts", breaks = range(df$most_sig)) + 
  scale_y_reverse() + 
  ggtitle(paste0(most_sig, " expression")) + 
  theme_bw() + 
  theme(aspect.ratio = 1, 
        panel.grid = element_blank(), 
        plot.title = element_text(face = "italic"), 
        axis.title = element_blank(), 
        axis.text = element_blank(), 
        axis.ticks = element_blank())


# ----------------------------------------------
# volcano plot: standard significance thresholds
# ----------------------------------------------

# summarize results with volcano plot

# standard significance thresholds

logfc <- res$coefficients[, "annot_region_pseudoLC"]

stopifnot(length(fdrs) == length(logfc))
stopifnot(all(fdrs_gene_ids == names(logfc)))

names(logfc) <- names(fdrs)

# identify significant genes (low FDR and high logFC)
thresh_fdr <- 0.05
thresh_logfc <- log2(2)
sig <- (fdrs < thresh_fdr) & (abs(logfc) > thresh_logfc)

# number of significant genes
table(sig)


df <- data.frame(
  gene = names(fdrs), 
  FDR = fdrs, 
  log2FC = logfc, 
  sig = sig
)

pal <- c("black", "red")


# volcano plot without labels
ggplot(df, aes(x = log2FC, y = -log10(FDR), color = sig)) + 
  geom_point(size = 0.1) + 
  geom_point(data = df[df$sig, ], size = 0.5) + 
  scale_color_manual(values = pal, guide = "none") + 
  geom_hline(yintercept = -log10(0.05), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = -log2(2), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = log2(2), lty = "dashed", color = "royalblue") + 
  ggtitle("LC vs. non-LC regions") + 
  theme_bw() + 
  theme(plot.title = element_text(face = "bold"), 
        panel.grid.minor = element_blank())

fn <- file.path(dir_plots, "pseudobulkDE_volcano_significant")
ggsave(paste0(fn, ".pdf"), width = 4.5, height = 4)
ggsave(paste0(fn, ".png"), width = 4.5, height = 4)


# -------------------------------------------
# volcano plot: highly significant thresholds
# -------------------------------------------

# summarize results with volcano plot

# highly significant thresholds

# identify significant genes (low FDR and high logFC)
thresh_fdr <- 1e-3
thresh_logfc <- log2(3)
highlysig <- (fdrs < thresh_fdr) & (abs(logfc) > thresh_logfc)

# number of significant genes
table(highlysig)


df <- data.frame(
  gene = names(fdrs), 
  FDR = fdrs, 
  log2FC = logfc, 
  highlysig = highlysig
)

pal <- c("black", "red")


# volcano plot without labels
ggplot(df, aes(x = log2FC, y = -log10(FDR), color = highlysig)) + 
  geom_point(size = 0.1) + 
  geom_point(data = df[df$highlysig, ], size = 0.5) + 
  scale_color_manual(values = pal, guide = "none") + 
  geom_hline(yintercept = -log10(1e-3), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = -log2(3), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = log2(3), lty = "dashed", color = "royalblue") + 
  ggtitle("LC vs. non-LC regions") + 
  theme_bw() + 
  theme(plot.title = element_text(face = "bold"), 
        panel.grid.minor = element_blank())

fn <- file.path(dir_plots, "pseudobulkDE_volcano_highlySig")
ggsave(paste0(fn, ".pdf"), width = 4.5, height = 4)
ggsave(paste0(fn, ".png"), width = 4.5, height = 4)


# volcano plot with labels
set.seed(123)
ggplot(df, aes(x = log2FC, y = -log10(FDR), color = highlysig, label = gene)) + 
  geom_point(size = 0.1) + 
  geom_point(data = df[df$highlysig, ], size = 0.5) + 
  geom_text_repel(data = df[df$highlysig, ], 
                  size = 1.5, nudge_y = 0.1, 
                  force = 0.1, force_pull = 0.1, min.segment.length = 0.1, 
                  max.overlaps = 20) + 
  scale_color_manual(values = pal, guide = "none") + 
  geom_hline(yintercept = -log10(1e-3), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = -log2(3), lty = "dashed", color = "royalblue") + 
  geom_vline(xintercept = log2(3), lty = "dashed", color = "royalblue") + 
  ggtitle("LC vs. non-LC regions") + 
  theme_bw() + 
  theme(plot.title = element_text(face = "bold"), 
        panel.grid.minor = element_blank())

fn <- file.path(dir_plots, "pseudobulkDE_volcanoWithLabels_highlySig")
ggsave(paste0(fn, ".pdf"), width = 4.5, height = 4)
ggsave(paste0(fn, ".png"), width = 4.5, height = 4)


# -------
# heatmap
# -------

stopifnot(nrow(mat) == nrow(spe_pseudo))
stopifnot(all(rownames(mat) == rowData(spe_pseudo)$gene_id))
rownames(mat) <- rowData(spe_pseudo)$gene_name


# calculate mean logcounts across samples in LC and non-LC regions
# where mean is calculated as unweighted mean across samples
mat_LC <- mat[, 9:16]
mat_non <- mat[, 1:8]

mean_LC <- rowMeans(mat_LC)
mean_non <- rowMeans(mat_non)

hmat <- cbind(
  LC = mean_LC, 
  `non-LC` = mean_non
)


# select top genes
# ordered by FDR within set of significant genes (highly significant set)
stopifnot(length(fdrs) == length(highlysig))
top <- sort(fdrs[highlysig])
top_names <- names(top)

# order matrix rows
hmat <- hmat[top_names, ]

# gene names in row names
rownames(hmat) <- names(top)

# genes to highlight
ix_known <- which(rownames(hmat) %in% c("DBH", "TH", "SLC6A2", "SLC18A2"))
fontfaces <- rep("italic", nrow(hmat))
fontfaces[ix_known] <- "bold.italic"
fontcolors <- rep("black", nrow(hmat))
fontcolors[ix_known] <- "red"

row_annot <- rowAnnotation(
  rows = anno_text(rownames(hmat), 
                   gp = gpar(fontface = fontfaces, col = fontcolors, fontsize = 9))
)


# create heatmap
hm <- Heatmap(
  hmat, 
  col = viridis(100), 
  cluster_rows = FALSE, cluster_columns = FALSE, 
  column_names_rot = 0, column_names_gp = gpar(fontsize = 10), column_names_centered = TRUE, 
  right_annotation = row_annot, show_row_names = FALSE, 
  #row_names_gp = gpar(fontsize = 9, fontface = "italic"), 
  column_title = "LC vs. non-LC regions", 
  column_title_gp = gpar(fontsize = 10, fontface = "bold"), 
  name = "mean\nlogcounts"
)

hm

# save heatmap
fn <- file.path(dir_plots, "pseudobulkDE_heatmap")

pdf(paste0(fn, ".pdf"), width = 3.1, height = 6.5)
hm
dev.off()

png(paste0(fn, ".png"), width = 3.1 * 200, height = 6.5 * 200, res = 200)
hm
dev.off()


# -------
# MA plot
# -------

df <- data.frame(
  gene = names(fdrs), 
  mean_LC = mean_LC, 
  mean_non = mean_non, 
  mean = (mean_LC + mean_non) / 2, 
  logFC = logfc, 
  FDR = fdrs, 
  sig = sig
)


pal <- c("black", "red")

ggplot(df, aes(x = mean, y = logFC, color = sig, label = gene)) + 
  geom_point(size = 0.1) + 
  geom_point(data = df[df$sig, ], size = 0.5) + 
  scale_color_manual(values = pal, guide = "none") + 
  geom_hline(yintercept = log2(2), lty = "dashed", color = "royalblue") + 
  geom_hline(yintercept = -log2(2), lty = "dashed", color = "royalblue") + 
  labs(x = "mean logcounts (pseudobulked LC and non-LC)", 
       y = "log fold change") + 
  ggtitle("Pseudobulked DE tests: LC vs. non-LC") + 
  theme_bw() + 
  theme(panel.grid.minor = element_blank())

fn <- file.path(dir_plots, "pseudobulkDE_MAplot")
ggsave(paste0(fn, ".pdf"), width = 4.25, height = 4)
ggsave(paste0(fn, ".png"), width = 4.25, height = 4)


# -----------------
# export gene lists
# -----------------

stopifnot(length(fdrs_gene_ids) == nrow(rowData(spe_pseudo)))
stopifnot(all(fdrs_gene_ids == rowData(spe_pseudo)$gene_id))

stopifnot(length(fdrs_gene_ids) == length(sig))
stopifnot(all(rowData(spe_pseudo)$gene_name == names(sig)))
stopifnot(length(fdrs_gene_ids) == length(highlysig))
stopifnot(all(rowData(spe_pseudo)$gene_name == names(highlysig)))


# all genes
df_all <- data.frame(
  gene_id = fdrs_gene_ids, 
  gene_name = fdrs_gene_names, 
  source = rowData(spe_pseudo)$source, 
  type = rowData(spe_pseudo)$type, 
  gene_version = rowData(spe_pseudo)$gene_version, 
  gene_type = rowData(spe_pseudo)$gene_type, 
  mean_logcounts_LC = mean_LC, 
  mean_logcounts_nonLC = mean_non, 
  log2FC = logfc, 
  p_value = p_vals, 
  FDR = fdrs, 
  significant = sig, 
  highly_significant = highlysig
)


# subset significant and highly significant genes
df_sig <- df_all[sig, ]
df_highlysig <- df_all[highlysig, ]

dim(df_all)
dim(df_sig)
dim(df_highlysig)


# order by FDRs
df_all <- df_all[order(df_all$FDR), ]
df_sig <- df_sig[order(df_sig$FDR), ]
df_highlysig <- df_highlysig[order(df_highlysig$FDR), ]


# save .csv files
fn_all <- file.path(dir_outputs, "LC_pseudobulkDE_all.csv")
write.csv(df_all, file = fn_all, row.names = FALSE)

fn_sig <- file.path(dir_outputs, "LC_pseudobulkDE_sigGenes.csv")
write.csv(df_sig, file = fn_sig, row.names = FALSE)

fn_highlysig <- file.path(dir_outputs, "LC_pseudobulkDE_highlySigGenes.csv")
write.csv(df_highlysig, file = fn_highlysig, row.names = FALSE)
